Restructure proxies and comment functions
This commit introduces generic proxying for nested indexing of
iterables, so that all properties and methods are exposed automatically.
Special-case proxies for style and classList are not required, and an
update method for computed setting is now possible.

The proxy handlers are separated from the functions that create the
proxies where possible, and additional methods are separated from the
proxies.